Rate throttler refactorings
