+++ client/src/test/scala/org/latestbit/slack/morphism/client/ratectrl/RateControlParamsTestsSuite.scala
@@ -16,10 +16,20 @@
  *
  */
 
-package org.latestbit.slack.morphism.client.ratectl.impl
+package org.latestbit.slack.morphism.client.ratectrl
 
-case class RateThrottlerWorkspaceMetrics(
-    wholeWorkspaceMetric: Option[RateThrottlerMetric],
-    tiers: Map[Int, RateThrottlerMetric],
-    updated: Long
-)
+import org.scalatest.flatspec.AnyFlatSpec
+
+import scala.concurrent.duration._
+
+class RateControlParamsTestsSuite extends AnyFlatSpec {
+
+  val params = RateControlParams(
+    globalMaxRateLimit = Some(
+      10,
+      1.second
+    )
+  )
+
+  assert( params.globalMaxRateLimit.exists( limit => limit.value == 10 && limit.per.toSeconds == 1 ) )
+}
